+++ Assets/scripts/battle stuff/battle ui/Indicator.cs
@@ -21,7 +21,7 @@ public class Indicator : MonoBehaviour
 
     public SpriteRenderer sprite;
 
-    float finalLerpStatus;
+    double finalLerpStatus;
 
     public bool isBar;
 
@@ -125,18 +125,18 @@ public class Indicator : MonoBehaviour
     {
         activated = BattleManager.current.battleStarted;
 
-        float lerpStatus = TrackTimeManager.songPositionInBeats / beatOfThisNote;
+        double lerpStatus = TrackTimeManager.songPositionInBeats / beatOfThisNote;
         //main travel lerp
         if (activated && lerpStatus < 1)
         {
-            transform.position = Vector3.Lerp(start, end, TrackTimeManager.songPositionInBeats / beatOfThisNote) + transform.parent.position;
+            transform.position = Vector3.Lerp(start, end, (float)TrackTimeManager.songPositionInBeats / beatOfThisNote) + transform.parent.position;
         }
         else if (activated && lerpStatus >= 1)
         {
             finalLerpStatus = TrackTimeManager.songPositionInBeats - beatOfThisNote;
             //so now what we're going to do is lerp for one more unit over one more bar
 
-            transform.position = Vector3.Lerp(end, end - new Vector3(0, end.y + 1, 0), finalLerpStatus) + transform.parent.position;
+            transform.position = Vector3.Lerp(end, end - new Vector3(0, end.y + 1, 0), (float)finalLerpStatus) + transform.parent.position;
         }
